backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
youtube-transcript-api
youtube-search-python
httpx<0.28
google-generativeai
pymongo
python-dotenv
thefuzz